Brought MakeArrays class into its own file, fixed tests
